% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_eaf_to_xml.R
\name{mc_build_xml}
\alias{mc_build_xml}
\title{Construct XML tree}
\usage{
mc_build_xml(mctext, vkey, skipempty = TRUE)
}
\arguments{
\item{eaffile}{Path to and filename of an EAF file to be converted.}
}
\value{
An XML tree object of classes \code{XMLInternalDocument} and
  \code{XMLAbstractDocument}, as defined by the
  \code{\link[XML:xmlTree]{XML}} package.
}
\description{
Uses the \code{\link[XML:xmlTree]{XML}} package to construct the XML tree.
}
\keyword{internal}
