% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_eaf_to_mcr.R
\name{mc_convert_mcr}
\alias{mc_convert_mcr}
\title{Shape input from EAF files}
\usage{
mc_convert_mcr(eaffile)
}
\arguments{
\item{eaffile}{Path to and filename of an EAF file to be converted.}
}
\value{
A \code{\link[data.table]{data.table}} containing the annotation
  values of a single EAF file.
}
\description{
Called by \code{\link{mc_eaf_to_mcr}} for each EAF file in
the list. While reading the EAFs themselves is handled by
\code{\link{mc_read_eaf}}, this function brings the output into
the proper shape.
}
\keyword{internal}
