% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_read_eaf.R
\name{mc_meta_eaf}
\alias{mc_meta_eaf}
\title{Read external file metadata}
\usage{
mc_meta_eaf(metadata = NULL)
}
\arguments{
\item{metadata}{A file.}
}
\value{
Nothing yet.
}
\description{
Reads extra metadata not included in the EAF files from an external source.
To be used by the EAF-to-XML converter. Not implemented yet.
}
\keyword{internal}
