% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_utilities.R
\name{mc_missarg}
\alias{mc_missarg}
\title{Test for empty arguments}
\usage{
mc_missarg(x)
}
\value{
A logical vector of length 1.
}
\description{
\code{mcmissarg} tests whether or not a value was specified as an argument to
a function. Unlike \code{\link[base]{missing}}, it also returns \code{FALSE}
with zero-length objects such as \code{c()} or \code{data.frame()}.
}
\seealso{
\code{\link{multicast}}
}
\keyword{internal}
