% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_table.R
\name{mc_prep_table}
\alias{mc_prep_table}
\title{Prepare a summary table}
\usage{
mc_prep_table(xtab, data, by, writeto, funcs, bycol)
}
\arguments{
\item{xtab}{A table generated by \code{\link{mc_table}}.}

\item{data}{The \code{data.table} from which the table was created.}

\item{by}{Create table for \code{"all"} data, or split by \code{"corpus"} or \code{"text"}.}

\item{writeto}{Path to write output to.}

\item{funcs}{List of function symbols to tabulate.}

\item{bycol}{Column to split data by.}
}
\value{
Nothing.
}
\description{
Called by \code{\link{mc_table}}. This is a wrapper for the loops that
generate tables collection-wide or corpus/text-wise.
}
\keyword{internal}
