% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_eaf_to_xml.R
\name{mc_prep_xml}
\alias{mc_prep_xml}
\title{Shape input from EAF files for XML output}
\usage{
mc_prep_xml(eaffile)
}
\arguments{
\item{eaffile}{Path to and filename of an EAF file to be converted.}
}
\value{
A \code{\link[data.table]{data.table}} containing the annotation
  values of a single EAF file.
}
\description{
Called by \code{\link{mc_eaf_to_xml}} for each EAF file in
the list. While reading the EAFs themselves is handled by
\code{\link{mc_read_eaf}}, this function brings the output into
its proper shape.
}
\keyword{internal}
