% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_read_eaf.R
\name{mc_read_eaf}
\alias{mc_read_eaf}
\title{Read EAF file content}
\usage{
mc_read_eaf(eaffile)
}
\arguments{
\item{eaffile}{Path to and filename of an EAF file to be converted.}
}
\value{
A \code{\link[data.table]{data.table}} containing all data fields of
  the EAF file.
}
\description{
Reads a specified EAF file. As EAF files are XML files, this function relies
on the \code{xml2} package and XPath to do its job. The EAF file to be read
must have a certain internal structure as dictated by the Multi-CAST design;
see the Multi-CAST documentation for more information.
}
\keyword{internal}
