% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_eaf_to_mcr.R
\name{mc_write_mcr}
\alias{mc_write_mcr}
\title{Write a multicastR table to file}
\usage{
mc_write_mcr(raw, writeto, filename)
}
\arguments{
\item{raw}{A \code{\link[data.table]{data.table}} in multicastR format.}

\item{writeto}{Directory to which to write output to. Defaults to
\code{getwd}.}

\item{filename}{A length 1 character vector containing the name of the
written output. If empty, defaults to "multicast_YYMM".}
}
\value{
None.
}
\description{
Writes a table generated by \code{\link{mc_convert_mcr}} to a \code{txt}
file.
}
\keyword{internal}
