% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_reflist.R
\name{mc_write_refs_tex}
\alias{mc_write_refs_tex}
\title{Write referent lists as TEX}
\usage{
mc_write_refs_tex(refdata, writeto)
}
\arguments{
\item{refdata}{A \code{data.table} containing a list of referents.}

\item{writeto}{A directory to which to write output. Defaults to
\code{getwd}.}
}
\value{
Nothing.
}
\description{
Called by \code{\link{mc_reflist}}. Prepares the TSV lists of referents by
adding LaTeX markup, then writes them to file.
}
\keyword{internal}
