% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_reflist.R
\name{mc_write_refs_tsv}
\alias{mc_write_refs_tsv}
\title{Write referent lists as TSV}
\usage{
mc_write_refs_tsv(refdata, writeto)
}
\arguments{
\item{refdata}{A \code{data.table} containing lists of referents.}

\item{writeto}{A directory to which to write output. Defaults to
\code{getwd}.}
}
\value{
Nothing.
}
\description{
Called by \code{\link{mc_reflist}}. Writes bundled lists of referents.
}
\keyword{internal}
