% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_table.R
\name{mc_write_table}
\alias{mc_write_table}
\title{Write summary tables to file}
\usage{
mc_write_table(ctab, data, by, writeto, funcs, bycol)
}
\arguments{
\item{ctab}{A table generated by \code{\link{mc_table}}, then prepared by
\code{\link{mc_prep_table}}.}

\item{data}{The \code{data.table} from which the table was created.}

\item{by}{Write files for \code{"all"} data, or split by \code{"corpus"} or
\code{"text"}.}

\item{writeto}{Path to write output to.}

\item{funcs}{List of function symbols to tabulate.}

\item{bycol}{Column to split data by.}
}
\value{
Nothing.
}
\description{
Called by \code{\link{mc_prep_table}}.
}
\keyword{internal}
