% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_eaf_to_tex.R
\name{mc_write_tex}
\alias{mc_write_tex}
\title{Generate a TeX'ed parallel text from an EAF and write it to file}
\usage{
mc_write_tex(texts, writeto)
}
\arguments{
\item{writeto}{Directory to which to write output to.}

\item{eaffile}{Path to and filename of an EAF file to be converted.}
}
\description{
Generates a TeX file from an EAF file. Make sure to eliminate or escape all
special characters considered active by TeX before conversion!
}
\keyword{internal}
