% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_eaf_to_tsv.R
\name{mc_write_tsv}
\alias{mc_write_tsv}
\title{Write a TSV table to file}
\usage{
mc_write_tsv(mctext, writeto, filename, split = FALSE,
  bycorpus = FALSE)
}
\arguments{
\item{writeto}{Directory to which to write output to. Defaults to
\code{getwd}.}

\item{filename}{A length 1 character vector containing the name of the
written output. If empty and \code{split} is \code{FALSE}, defaults to
"multicast_YYMM". Ignored if \code{split} is \code{TRUE}, as file names are
automatically constructed.}

\item{split}{Logical. If \code{TRUE}, generates file names from text
metadata.}

\item{bycorpus}{Logical. If \code{TRUE}, generates file names only from
corpus names.}

\item{mcctext}{A \code{\link[data.table]{data.table}}.}
}
\description{
Writes a table generated by \code{\link{mc_prep_tsv}} to a TSV file.
}
\keyword{internal}
