% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_eaf_to_xml.R
\name{mc_write_xml}
\alias{mc_write_xml}
\title{Write an XML table to file}
\usage{
mc_write_xml(mctext, writeto, filename)
}
\arguments{
\item{mctext}{An XML document generated with \code{\link[XML:xmlTree]{XML}}.}

\item{writeto}{Directory to which to write output to. Defaults to
\code{getwd}.}

\item{filename}{A length 1 character vector containing the name of the
written output.}
}
\description{
Writes an XML tree generated by \code{\link{mc_prep_xml}} to an XML file.
}
\keyword{internal}
