% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_index.R
\name{mc_index}
\alias{mc_index}
\title{Access the Multi-CAST version index}
\usage{
mc_index()
}
\value{
A \code{\link{data.frame}} with five columns: \describe{
  \item{\code{[, 1] version}}{Version key. Used for the \code{vkey} argument
  of other functions in this package.} \item{\code{[, 2] date}}{Publication
  date in YYYY-MM-DD format.} \item{\code{[, 3] corpora}}{Number of corpora
  (languages).} \item{\code{[, 4] texts}}{Number of texts.} \item{\code{[, 5]
  size}}{Total file size in kilobytes.} }
}
\description{
\code{mc_index} downloads a tabular index of the versions of the Multi-CAST
corpus data from the servers of the University of Bamberg. The value in the
leftmost \code{version} column may be passed to the \code{\link{multicast}}
method for access to earlier versions of the annotations.
}
\examples{
\dontrun{
  # retrieve version index
  mc_index()
}

}
\seealso{
\code{\link{multicast}}, \code{\link{mc_metadata}},
  \code{\link{mc_referents}}, \code{\link{mc_clauses}}
}
