% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_read_eaf.R
\name{mc_meta_eaf}
\alias{mc_meta_eaf}
\title{Read EAF file metadata}
\usage{
mc_meta_eaf(raweaf)
}
\arguments{
\item{raweaf}{An XML file read by \code{xml2::read_xml}.}
}
\value{
Nothing yet.
}
\description{
To be used by the EAF-to-XML converter. Not used yet.
}
\keyword{internal}
