% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multicast.R
\name{multicast}
\alias{multicast}
\title{Access Multi-CAST annotation data}
\usage{
multicast(vkey, legacy.colnames = FALSE)
}
\arguments{
\item{vkey}{A numeric or character vector of length 1 specifying the
requested version of the annotation values. Must be one of the four-digit
version keys in the first column of \code{\link{mc_index}}, or empty. If
empty or no value is supplied, \code{multicast} automatically retrieves the
most recent version of the annotations. See the examples below for an
illustration.}

\item{legacy.colnames}{If \code{TRUE}, renames the \code{text} and
\code{gword} columns to what they were called prior to version 1.1.0 of the
package (i.e. \code{file}, \code{word}). This option will be removed in the
future.}
}
\value{
A \code{\link[data.table]{data.table}} with eleven columns:
  \describe{ \item{\code{[, 1] corpus}}{The name of the corpus.}
  \item{\code{[, 2] text}}{The title of the text. If \code{legacy.colnames}
  is \code{TRUE}, this column is named \code{file} instead.} \item{\code{[,
  3] uid}}{The utterance identifier. Uniquely identifies an utterance within
  a text.} \item{\code{[, 4] gword}}{Grammatical words. The tokenized
  utterances in the object language. If \code{legacy.colnames} is
  \code{TRUE}, this column is named \code{word} instead.} \item{\code{[, 5]
  gloss}}{Morphological glosses following the Leipzig Glossing Rules.}
  \item{\code{[, 6] graid}}{Annotations using the GRAID scheme (Haig &
  Schnell 2014).} \item{\code{[, 7] gform}}{The form symbol of a GRAID
  gloss.} \item{\code{[, 8] ganim}}{The person-animacy symbol of a GRAID
  gloss.} \item{\code{[, 9] gfunc}}{The function symbol of a GRAID gloss.}
  \item{\code{[, 10] refind}}{Referent tracking using the RefIND scheme
  (Schiborr et al. 2018).} \item{\code{[, 11] reflex}}{The information status
  of newly introduced referents, using a simplified version of the RefLex
  scheme (Riester & Baumann 2017).} }
}
\description{
\code{multicast} downloads the Multi-CAST annotation data from the servers of
the University of Bamberg and outputs them as a
\code{\link[data.table]{data.table}}. As the Multi-CAST collection is
amenable to extension by additional data sets and annotation schemes,
\code{multicast} takes an optional argument to select earlier versions of the
annotation data to ensure scientific accountability and reproducability.
}
\section{Licensing}{
 The Multi-CAST annotation data accessed by the
  \code{multicast} method is published under a \emph{Create Commons
  Attribution 4.0 International} (CC-BY 4.0) licence
  (\url{https://creativecommons.org/licenses/by-sa/4.0/}). Please refer to
  the collection documentation for information on how to give proper credit
  to its contributors.
}

\section{Citing Multi-CAST}{
 Data from the Multi-CAST collection should be
  cited as: \itemize{ \item Haig, Geoffrey & Schnell, Stefan (eds.).
  2015. \emph{Multi-CAST: Multilinguial Corpus of Annotated Spoken
  Texts}. (\url{https://multicast.aspra.uni-bamberg.de/}) (Accessed
  \emph{date}.) } If for some reason you need to cite this package on its
  own, please refer to \code{citation(multicastR)}.
}

\section{References}{
 \itemize{\item Haig, Geoffrey & Schnell, Stefan. 2014.
  \emph{Annotations using GRAID (Grammatical Relations and Animacy in
  Discourse): Introduction and guidelines for annotators.} Version 7.0.
  (\url{https://multicast.aspra.uni-bamberg.de/#annotations})

  \item Riester, Arndt & Baumann, Stefan. 2017. The RefLex scheme --
  Annotation guidelines. \emph{SinSpeC: Working papers of the SFB 732} 14.
  (\url{https://dx.doi.org/10.18419/opus-9011}))

  \item Schiborr, Nils N. & Schnell, Stefan & Thiele, Hanna. 2018.
  \emph{RefIND -- Referent Indexing in Natural-language Discourse: Annotation
  guidelines.} Version 1.1.
  (\url{https://multicast.aspra.uni-bamberg.de/#annotations})}
}

\examples{
\dontrun{
  # retrieve and print the most recent version of the
  # Multi-CAST annotations
  multicast()

  # retrieve and print the version of the annotation data
  # published in May 2019
  multicast(1905)   # or: multicast("1905")
}

}
\seealso{
\code{\link{mc_index}}, \code{\link{mc_referents}}
}
